Gamepad moved to a separate namespace, update public property name
